+++ include/primitive_lists.h
@@ -1,0 +1,13 @@
+#ifndef __CLIB_PRIMITIVE_LISTS_H__
+#define __CLIB_PRIMITIVE_LISTS_H__
+
+#define TYPE int
+#define COPY_VALUE 
+#include "list.h"
+
+#define TYPE double
+#define COPY_VALUE
+#include "list.h"
+
+#endif
+
