+++ test/list.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include "list.h"
+#include "primitive_lists.h"
 
 int main()
 {
